refactor(tests): delete inference, safety and agents tests from providers/tests/
