+++ tests/integration/test_cases/inference/chat_completion.json
