+++ tests/integration/test_cases/inference/completion.json
